Refactor rename: settings now live under IDE
