+++ pyiron_base/ide/settings.py
@@ -30,7 +30,7 @@ import os
 import importlib
 from configparser import ConfigParser
 from pyiron_base.generic.singleton import Singleton  # Ok, this is the one exception to not importing from pyiron...
-from pyiron_base.settings.logger import setup_logger
+from pyiron_base.ide.logger import setup_logger
 
 __author__ = "Jan Janssen"
 __copyright__ = (
